fix artifact paths
